full open source include PCB

+++ QueFOC_CH32V307/Hardware/it.c
@@ -79,7 +79,7 @@ void UART4_IRQHandler(void)
     if(USART_GetITStatus(UART4, USART_IT_RXNE) != RESET)
     {
         u16 data = USART_ReceiveData(UART4);
-        rec_data_package_float(data, hardware_uart4_callback);
+        rec_data_package_float((u8)data, hardware_uart4_callback);
         USART_ClearITPendingBit(UART4, USART_IT_RXNE);
     }
 }
